requirements.txt
openai==0.28.0  # Versão específica para compatibilidade
anthropic==0.7.7
python-dotenv==1.0.0  # Para variáveis de ambiente
PyMuPDF==1.23.8  # Para processamento de PDFs
pdf2image==1.16.3  # Para OCR
pytesseract==0.3.10  # Para OCR
Pillow==10.1.0  # Para processamento de imagens
customtkinter==5.2.1  # Para interface gráfica moderna
PyYAML  # Para arquivos de configuração
requests==2.31.0
python-docx==1.0.1
tqdm==4.66.1
tiktoken==0.5.1
httpx>=0.25.0  # Para suporte a proxy
nltk>=3.8.1
unidecode>=1.3.8
pyyaml>=6.0.1
python-dateutil>=2.8.2
